9/2/13 home
Got the open and close operations on the binary save files working.
Haven't filly tested. compiles and runs fine and seems to load the blank
info into the controls properly. Will need to finish creating the
commander edit controls and allow it to save new items into the vector.
Also need to search the vector to see if the commander exists already
and overwrite.