Changed to unicode
Converted the project to unicode using WCHAR in place of char in most
places. This broke the file I/O with iStream and fStream which require
char[]'s instead of the WCHAR supplied wchar_t's or whatever. Need to
convert from istream and fstream file I/O to windows file management.

See for reference:

http://msdn.microsoft.com/en-us/library/windows/desktop/bb540534(v=vs.85).aspx